WIP: Use spnego to loop in basic_auth
This requires that the gethostname() returns a host name that can
be used to acquire a ticket for HTTP@<gethostname()>.
If the keytab has principal names that do not conform validation
will fail.

NOTE: The problem with this implementation is that passing gss_mech_spnego
to acquire_creds_wth_pasword() causing it to run over the list of mechanisms
and perform thre different AS requests for the three available krb5-related
oids even if we will end up calling gss_set_neg_mechs() to restrict the
mechanism used to a completely different mech. This has a way to high impact
to be worthwhile for now given also the name limitation.

Signed-off-by: Simo Sorce <simo@redhat.com>